Use `assets` instead of `chunks`.
There are some cases when assets are generated without chunks, e.g. using `file-loader`. In those cases the previous result of `webpack-assets` did not correctly include them. Now it reads `stats.assets` for the information. It should be kind of API compatible, though the `hash` calculation is different, `name` is now the complete name (what `file` was before) and the input requests `stats.assets` to be present.